+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [1.0.14]
+### Added
+* Added possibility to configure the transformer in order to set the default value for all destination's object fields that are not existing in the source object.
+  See [README.md](https://github.com/HotelsDotCom/bull/blob/master/README.md) for more details.
+### Changed
+* Jumped to version 1.0.14 in order to be consequent to the previous library version hosted on a private repo.
+
 ## [1.0.3]
 ### Added
 * Added changelog file.